Fix undefined being output for some generated attributes
This happens if the API responded with an attribute key in the
"template" response but no associated value in the "attribute values"
response. (potential bigger change: have the API respond directly with
key-value pairs instead of separate rows for "keys" and "values").

Both the "template" and the "generated title" have now been corrected to
remove the empty-value attribute appearing in the "template" and
"undefined" appearing in the title.

This change also includes removing unnecessary logging and writes error
output into the Full API response column (as the "log" sheet is hidden
by default, and it's difficult to find the corresponding entry as we
don't log with any row information like Item ID for example).