Remove filling of non-gaps with original data for identified gap columns
The default settings of Merchant Center supplemental feeds are to use
values from the Main feed if they're not present in the supplemental
feed. Accordingly, there is no need for us to prefill non-gaps with the
original data anymore (this is also better for UX as users now just see
which values were filled in the Output Feed worksheet).